delta project/requirements.txt
blinker==1.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
delta-spark==2.3.0
distro==1.7.0
distro-info===1.1build1
exceptiongroup==1.1.1
httplib2==0.20.2
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
joblib==1.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
packaging==23.1
pluggy==1.0.0
py4j==0.10.9.5
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyspark==3.3.2
pytest==7.3.1
python-apt==2.4.0+ubuntu1
PyYAML==5.4.1
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
threadpoolctl==3.1.0
tomli==2.0.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
zipp==1.0.0
